+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
 ### Added
+
 - `LoggingConfig.handlers` for configuring logging handlers.
 - `LoggingConfig.handlers_formatter` for configuring a formatter for all handlers.
 - `LoggingConfig.propagate` for configuring whether log messages should be propagated to parent loggers.
@@ -15,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Request/response debug logging to all Prompt Drivers.
 
 ### Changed
+
 - `_DefaultsConfig.logging_config` and `Defaults.drivers_config` are now lazily instantiated.
 
 ## \[0.33.0\] - 2024-10-09